Stabilize frame rate
Before, a drop in recording frame rate would result in the output skipping the missed frames and making the output seem to speed up.

Now, the recorder tracks the timeSinceRecordingStart and passes it on to the GifConverter to account for it when creating the gif. It's done by changing the frame delay for every individual frame.